Update badge image
